Move domain specs into domain directory
