+++ src/tests/domain/formatters/duration-formatter.spec.ts
@@ -15,7 +15,7 @@
  * along with this program.  If not, see <http://www.gnu.org/licenses/>.
  */
 
-import { DurationFormatter } from '../../domain/formatters/duration-formatter'
+import { DurationFormatter } from '../../../domain/formatters/duration-formatter'
 
 describe('DurationFormatter', () => {
   const formatter = new DurationFormatter()
